requirements.txt
# Backend dependencies only (Streamlit replaced by React frontend)
sentence-transformers>=3.0.0
fastapi==0.111.0
uvicorn==0.30.1
python-dotenv==1.0.1
pydantic>=2.10.0
requests==2.32.3
langchain==0.2.6
langchain-community==0.2.6
langchain-groq==0.1.5
langchain-huggingface==0.0.3
faiss-cpu>=1.8.0
